devp2p: migrate additional types to src/types.ts
