+++ packages/devp2p/src/dpt/dpt.ts
@@ -9,75 +9,10 @@ import { BanList } from './ban-list.js'
 import { KBucket } from './kbucket.js'
 import { Server as DPTServer } from './server.js'
 
-import type { PeerInfo } from '../types.js'
+import type { DPTOptions, PeerInfo } from '../types.js'
 import type { Debugger } from 'debug'
 
 const DEBUG_BASE_NAME = 'dpt'
-
-interface DPTOptions {
-  /**
-   * Timeout for peer requests
-   *
-   * Default: 10s
-   */
-  timeout?: number
-
-  /**
-   * Network info to send a long a request
-   *
-   * Default: 0.0.0.0, no UDP or TCP port provided
-   */
-  endpoint?: PeerInfo
-
-  /**
-   * Function for socket creation
-   *
-   * Default: dgram-created socket
-   */
-  createSocket?: Function
-
-  /**
-   * Interval for peer table refresh
-   *
-   * Default: 60s
-   */
-  refreshInterval?: number
-
-  /**
-   * Toggles whether or not peers should be queried with 'findNeighbours'
-   * to discover more peers
-   *
-   * Default: true
-   */
-  shouldFindNeighbours?: boolean
-
-  /**
-   * Toggles whether or not peers should be discovered by querying EIP-1459 DNS lists
-   *
-   * Default: false
-   */
-  shouldGetDnsPeers?: boolean
-
-  /**
-   * Max number of candidate peers to retrieve from DNS records when
-   * attempting to discover new nodes
-   *
-   * Default: 25
-   */
-  dnsRefreshQuantity?: number
-
-  /**
-   * EIP-1459 ENR tree urls to query for peer discovery
-   *
-   * Default: (network dependent)
-   */
-  dnsNetworks?: string[]
-
-  /**
-   * DNS server to query DNS TXT records from for peer discovery
-   */
-  dnsAddr?: string
-}
 
 export class DPT extends EventEmitter {
   privateKey: Uint8Array
